[process][riscv] Share more code with x86
Functions shared:

- switch_to_task_pop_nested_interrupts()
- adjust_nested_interrupts_for_task_in_kernel()
- save_current_task_state()
- save_regs_on_user_stack()
- restore_regs_from_user_stack()
- setup_pause_trampoline()
