+++ include/tilck/kernel/arch/i386/arch_utils.h
@@ -78,6 +78,11 @@ static ALWAYS_INLINE void *regs_get_ip(regs_t *r)
    return TO_PTR(r->eip);
 }
 
+static ALWAYS_INLINE void regs_set_ip(regs_t *r, ulong value)
+{
+   r->eip = value;
+}
+
 static ALWAYS_INLINE ulong regs_get_usersp(regs_t *r)
 {
    return r->useresp;
